Another explicitly-int length arg we can do nothing about
git-svn-id: https://svn.eu.apache.org/repos/asf/apr/apr-util/trunk@587435 13f79535-47bb-0310-9956-ffa450edef68

+++ crypto/apr_md5.c
@@ -709,7 +709,7 @@ APU_DECLARE(apr_status_t) apr_password_validate(const char *passwd,
         apr_md5_encode(passwd, hash, sample, sizeof(sample));
     }
     else if (!strncmp(hash, APR_SHA1PW_ID, APR_SHA1PW_IDLEN)) {
-         apr_sha1_base64(passwd, strlen(passwd), sample);
+         apr_sha1_base64(passwd, (int)strlen(passwd), sample);
     }
     else {
         /*
